requirements.txt
torch
torchvision
tqdm
wandb
python-dotenv